Поправил миграцию и табличку.

+++ src/migrations/versions/5a46f5026091_first_migration.py
@@ -54,7 +54,7 @@ def upgrade_write():
     )
     op.create_table('user',
     sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.Column('user_id', sa.BigInteger(), nullable=False),
     sa.Column('first_name', sa.String(length=64), nullable=True),
     sa.Column('last_name', sa.String(length=64), nullable=True),
     sa.Column('username', sa.String(length=64), nullable=True),
